Document attempt archive context

+++ admin/code/tce_attempt_archive.php
@@ -1,7 +1,8 @@
 <?php
 
 require_once '../config/tce_config.php';
-$pagelevel = K_AUTH_ADMIN_RESULTS;
+/** @var mixed $db Database connection initialized by tce_config.php. */
+$pagelevel = (int) K_AUTH_ADMIN_RESULTS;
 require_once '../../shared/code/tce_authorization.php';
 require_once '../../shared/code/tce_functions_attachments.php';
